Working sample

+++ src/main/java/com/qe/core/JobConfig.java
@@ -2,5 +2,5 @@ package com.qe.core;
 
 public interface JobConfig {
 
-    int getTransactionsPerSecond();
+    double getTransactionsPerSecond();
 }
